finalizando filtragem dos indicadores

+++ server/src/main/java/com/vincula/dto/projection/StatusQuantidadeProjection.java
@@ -2,5 +2,5 @@ package com.vincula.dto.projection;
 
 public interface StatusQuantidadeProjection {
     String getStatus();
-    Double getQuantidade();
+    int getQuantidade();
 }
